synchronizer: use floats for exponential timeout

+++ cmd/hotstuff/main.go
@@ -43,7 +43,7 @@ type options struct {
 	RootCAs         []string    `mapstructure:"root-cas"`
 	SelfID          hotstuff.ID `mapstructure:"self-id"`
 	TLS             bool
-	ViewTimeout     uint `mapstructure:"view-timeout"`
+	ViewTimeout     float64 `mapstructure:"view-timeout"`
 	Replicas        []replica
 }
